Make PluginLoader.LoadPlugin return IEnumerable<T>
With an eye towards a 1.0 release, changing the signature of this method
to a less restrictive return type gives us more freedom to change the
implementation without it becoming a breaking change. The previous
return type of ICollection<T> was also a bad semantic fit for this method,
since the mutating methods (.Add(), .Clear(), etc.) that it exposes make
no sense in this context.